+++ src/components/ProductList/ProductList.tsx
@@ -9,7 +9,7 @@ interface ProductListProps {
 }
 
 const ProductList: FC<ProductListProps> = ({ products }) => {
-
+    
     return (
         products.length > 0 ? (
             <div className={styles.container}>
