+++ src/index.tsx
@@ -1,11 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { setupStore } from './store/store';
 import { ThemeProvider } from '@mui/material';
 import { theme } from './helpers/theme';
+import './assets/styles/globals.css';
+import './assets/styles/reset.css';
+import './assets/styles/variables.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 const store = setupStore();
